+++ packages/orq-rc/RELEASES.md
@@ -8599,3 +8599,13 @@ Based on:
 - [typescript v4.8.0-rc.53] packages/orq-rc
 ### Releases
 - [NPM v4.8.0-rc.53] https://www.npmjs.com/package/@orq-ai/node/v/4.8.0-rc.53 - packages/orq-rc
+
+## 2026-04-23 13:30:32
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.9 (2.881.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.8.0-rc.54] packages/orq-rc
+### Releases
+- [NPM v4.8.0-rc.54] https://www.npmjs.com/package/@orq-ai/node/v/4.8.0-rc.54 - packages/orq-rc